Add an option to record hotmap
This saves records of all read request's addresses to /tmp/tgt_hotmap.

Data from this can later be used to visualize how much data is
accessed frequently.

The rationale behind this feature is to make it possible to
cache(via mlock(2)) specific ranges of a target image to speed-up boot
and launch of specific programs and reduce load of the backing-storage
device.

Any write requests will mark that address invalid (-1) as it's meaningless
to cache it as this is for speeding up read requests.

This feature can record up-to 127 accesses.
Any subsequent reads in that address won't increase the counter.

Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>