Redirect reads to the master image
Linux doesn't share page caches with CoW'ed files.

This leads to page faults everytime when the client reads the CoW'ed
slave image even when the same range has been cached in the master image.

To fix this, redirect reads to the master image if the client didn't
write to the given address.

This is implemented by maintaining a flag mapping table for each clients.

When a client writes data, the said client's flag mapping table is
flipped from 0 to 1 at the given address.

Subsequent read requests will be done with the client's own image, rather
than with the master image.

This flag mapping table is deliberately super large to remove potential
bitmasking overhead. This can technically be reduced to 1/8th the size
with bitmasking.

Now that the flag mapping table is big anyways, 1 GiB of memory is used to
maintain all flag mapping tables. Hugepage is used to remove TLB overhead.
This allows us to serve 100 clients with a 40 GiB image, or 68 clients
with a 60 GiB image.

Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>